Clean up object rendering
